+++ components/AuthComponents/LoginComponents/LoginForm.tsx
@@ -117,7 +117,7 @@ export const LoginForm = () => {
                 sx={{ marginBottom: 4, marginTop: 1 }}
               >
                 <LinkButton
-                  link="/PasswordReset"
+                  link="/Auth/PasswordReset"
                   text="forgot your password?"
                 />
               </Grid>
@@ -134,7 +134,7 @@ export const LoginForm = () => {
           </form>
           <p>
             <a style={{ paddingRight: 5 }}>don&apos;t have an account?</a>
-            <LinkButton link="/Register" text="sign up" />
+            <LinkButton link="/Auth/Register" text="sign up" />
           </p>
         </Grid>
         <Divider role="log in with google or facebook accounts">
